* main:
  lib/model: Fix race in GlobalDirectory tests (fixes #7021) (#7022)
  lib: Use uint64 for disk stats (ref #3930) (#7019)
  lib/model, lib/protocol: Send ClusterConfig on config change (fixes #7020) (#7018)
  all: Move remaining protos to use the vanity plugin (#7009)
  lib/config, lib/syncthing: Only drop delta index on upgrade if so ordered (fixes #6982) (#6983)
  gui, man, authors: Update docs, translations, and contributors
  lib: Remove HelloResult type, same as Hello (#7015)

+++ lib/scanner/walk.go
@@ -596,7 +596,7 @@ func CreateFileInfo(fi fs.FileInfo, name string, filesystem fs.Filesystem) (prot
 	}
 	f.Permissions = uint32(fi.Mode() & fs.ModePerm)
 	f.ModifiedS = fi.ModTime().Unix()
-	f.ModifiedNs = int32(fi.ModTime().Nanosecond())
+	f.ModifiedNs = fi.ModTime().Nanosecond()
 	if fi.IsDir() {
 		f.Type = protocol.FileInfoTypeDirectory
 		return f, nil
